bot/requirements.txt
# CtrlAltStock Blog Bot — Python Dependencies
# Install with: pip install -r requirements.txt

# Reddit API
praw==7.7.1

# RSS Feed parsing
feedparser==6.0.11

# HTTP client
httpx==0.27.0
beautifulsoup4==4.12.3

# AI providers
openai>=1.35.0
# ollama is accessed via httpx (REST) - no package needed

# Amazon PA-API SDK — NOT on PyPI, install manually if needed:
#   Download from: https://webservices.amazon.com/paapi5/documentation/quick-start/using-sdk.html
#   Or skip it — the bot uses Amazon search URLs as a fallback (still earns commission)
# paapi5-python-sdk==1.0.0

# Image handling (use >=11 for Python 3.13 support)
Pillow>=11.0.0

# Markdown utilities
python-frontmatter==1.1.0

# Scheduling
schedule==1.2.2

# Git operations
gitpython==3.1.43

# Environment variables
python-dotenv==1.0.1

# Database (deduplication)
# sqlite3 is built-in to Python, no install needed

# Retry logic
tenacity==8.3.0

# Utilities
python-slugify==8.0.4

# Web scraping fallback
lxml>=5.2.2
